Move hooks into new hook directory
